Update to ecstasy 2.0
Ecstasy 2.0 fixes a lot of the performance issues we were seeing! Woo!

It also breaks a bunch of the api. Boo!

This is the bare minimum to get the game working again with ecstasy 2.0. Further
work is needed to make things nice and take more advantage of the useful
functions introduced in the new update.